[Fix/#308]: 시큐리티 필터 파싱 오류 해결 (#309)

+++ api/src/main/kotlin/com/few/api/security/authentication/token/TokenUserDetailsService.kt
@@ -31,8 +31,9 @@ class TokenUserDetailsService(
 
         val id = claims.get(
             MEMBER_ID_CLAIM_KEY,
-            Long::class.java
-        )
+            Integer::class.java
+        ).toLong()
+
         val roles = claims.get(
             MEMBER_ROLE_CLAIM_KEY,
             String::class.java
